feat: log messages from middleware to sentry
Actix currently discards any errors generated in the middleware. You are
supposed to pass them in the body or some similar sub-optimal means.

Not a super fan of this approach, but basically, this patch attaches a
the error to the request or response extensions bucket. Since sentry is
also a middleware and gets run on every response, it sweeps the
request/response extension buckets and reports any errors it finds.

I created several new `ApiErrorKind`s to encapsulate the errors being
generated by the middleware layers. Partly to help isolate them and
partly to help any future filtering we may want to do.

Closes #504